Move overlay tests
