+++ packages/core/test/index.ts
@@ -57,7 +57,7 @@ import "./menu/menuTests";
 import "./multistep-dialog/multistepDialogTests";
 import "./non-ideal-state/nonIdealStateTests";
 import "./overflow-list/overflowListTests";
-import "./overlay2/overlayTests";
+import "./overlay/overlayTests";
 import "./panel-stack2/panelStack2Tests";
 import "./popover/popoverTests";
 import "./popover/popperUtilTests";
